+++ src/main/java/com/revature/organization/model/Organization.java
@@ -18,7 +18,7 @@ public class Organization {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column
-	private Integer id;
+	private Long id;
 	@Column(name="`name`")
 	private String name;
 	@Column(name="`alias_name`")
@@ -33,11 +33,13 @@ public class Organization {
 	private Boolean isActive=true;
 	
 
-	public Integer getId() {
+	
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
