+++ src/main/java/com/revature/organization/service/OrganizationService.java
@@ -9,9 +9,9 @@ public interface OrganizationService {
 	
 	List<Organization> get();
 	
-	Organization get(int id);
+	Organization get(Long id);
 	
 	void save(Organization org);
 	
-	void delete(int id);
+	void delete(Long id);
 }
